Vim: set low ttimeoutlen
MacVim isn't affected but "vim" in the Terminal is: typing a command
like "O" will result in a noticeable delay because it might be the first
character of a multibyte keycode (an arrow key, for instance).

This can be observed in the form of a noticeable delay between pressing
"O" and actually entering Insert mode. By default 'ttimeoutlen' is -1
and that means that default 'timeoutlen' value of 1000 milliseconds is
used for the detection of multibyte keycodes (ie. there is a 1 second
delay between pressing the "O" key and actually entering Insert mode).

Set 'ttimeoutlen' to a fast value (50 milliseconds) to make the delay
less noticeable. This should be fast enough to not cause any problems,
even when working over an SSH connection with some latency.

Signed-off-by: Wincent Colaiuta <win@wincent.com>