Vim: map Command-A, Command-O and Command-T
Unmap Command-A from the "Select All" and use it to jump to the
end of the line and start inserting (when already in Insert mode); this
is quicker than doing a <CTRL-o> followed by "A".

Relatedly, make Command-I jump to the beginning of the line (the first
non-blank position) and start inserting (when already in Insert mode);
again this is quicker than doing a <CTRL-o> followed by "I".

Similarly, make Command-0 jump to the absolute beginning of the line
(column zero) and start inserting (when already in Insert mode); this is
quicker than doing a <CTRL-o> followed by "0".

Unmap Command-O from "Open File..." and use it to open a new line below
the current and await in Insert mode (when already in Insert mode);
this is again quicker than doing a <CTRL-o> followed by "o".

As a companion, make Shift-Command-O open a new line above the current
one; this is quicker than doing a <CTRL-o> followed by "O".i

Finally, instead of using ,t to bring up fuzzyfinder, unmap Command-T
from "New Tab" and use it instead. This has the advantage of working in
all modes, not just Normal mode.

In making this change we unmap Shift-Command-T from "Open Tab..." and
reassign it to "New Tab" so that we continue to have a keyboard shortcut
for conveniently creating a new tab (although admittedly, :tabnew is not
too hard to type in any case).

Note that these mappings don't have any effect in command-line Vim, only
in MacVim, because the Terminal effectively intercepts all such key
combinations and Vim never "sees" them.

This is why we allow the old ,t mapping to continue on in the .vimrc
file, so that we can still bring up fuzzyfinder quickly when working
inside the Terminal.

Signed-off-by: Wincent Colaiuta <win@wincent.com>