Bash: change from GPL to BSD license
The GPL license header was left over from the days when the "bash
completion" stuff by Ian Macdonald was incorporated in the .bash_profile
file itself, which in turn meant that the entire file had to be licensed
under the GPL.

Now the bash completion stuff is not incorporated but rather referenced
(sourced), so the license can go. Rather than just ripping out the
license I am opting to apply the BSD license (the simplified, 2-clause
version as used by FreeBSD).

Signed-off-by: Wincent Colaiuta <win@wincent.com>