Bash: add JRuby bin directory to PATH
So that we can use Celerity/Culerity from within Cucumber/Capybara. We
add it to the end of the PATH so that items like JRuby's "gem" and "irb"
won't take precedence inappropriately over the MRI-based versions. To
actually run the JRuby version, do "jruby -S gem" or "jruby -S irb".

Signed-off-by: Wincent Colaiuta <win@wincent.com>