vim: use a more subtle, low-contrast color for NonText
This is a color that is applied on basically every single line (it is
used for the eol marker) so we want it to be something that is not too
jarring.

This is in contrast to the SpecialKeys highlighting, which in most cases
indicates the presence of an undesirable character, so in that case we
choose a relatively high-contrast color (blue).

Signed-off-by: Wincent Colaiuta <win@wincent.com>