Vim: drop <Leader>bda shortcut
Buffer navigation is way easier with the new Command-T, so no need to
clear out old buffers anymore with this shortcut.

The benefit of dropping it is it allows us to use <Leader>b for buffer
navigation without any pause (otherwise Vim has to wait after the 'b' to
see if the user is going to continue and type 'bda').

Signed-off-by: Wincent Colaiuta <win@wincent.com>