Vim: drop MiniBufExplorer
I'm getting a bit annoyed lately with a couple of glitches in
MiniBufExplorer (namely how it misresizes sometimes when closing and
splitting windows), and also how much space it starts to take up at the
end of a long editing session.

So drop it for now and see how life is without it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>