Git: remove user.name and user.email
I use a different email address on work and personal projects, so rather
than having a default which will be wrong for some projects, remove the
default and require that I set up user.name and user.email on a
per-project basis.

(I'd rather have to do some manual config for each project than push an
immutable commit with the wrong address in it for all eternity).

At least, will give it a try this way and see how it goes.