Bash: use $HOME/bin/vim as EDITOR, if available
Git is evidently using an emaciated PATH that doesn't include $HOME/bin
when searching for the EDITOR, which on some machines, leads to an old
and crusty Vim being used which doesn't know how to recognize the
"gitcommit" filetype.

Get around that by supplying the full path to the desired version of
Vim, if available.

Signed-off-by: Wincent Colaiuta <win@wincent.com>