Git: force standard less behavior
Recently I've been exporting LESS as MRX, whereas I used to use FMRX,
because I often want to be able to highlight all uses of a variable
within a file, even a short file, and I can't do that if less exits
immediately due to the -F option.

That's very annoying, however, when you use "git diff" and have to
explicitly exit the pager when there's empty output (eg. a clean
worktree). This forces Git to use its default behavior (FSRX) when it
uses the pager, instead of respecting the LESS environment variable.

Signed-off-by: Wincent Colaiuta <win@wincent.com>