Vim: make more Command-T mappings work in the terminal
These changes make both the ESC and the up and down cursor keys work in
the following places:

  - Mac OS X terminal, xterm-256color (iTerm)
  - Linux (CentOS terminal), xterm-256color via SSH
  - Linux (CentOS terminal) inside tmux, xterm-256color via SSH

While preserving the existing behavior in MacVim.

Oddly, the left and right cursor keys work everywhere without special
treatment.

Note that since I added "stty -ixon" to my .bash_profile <C-s> also
works in tmux. The only thing that works in MacVim which doens't work in
the terminal is <C-CR> to open in a horizontal split. (It seems to be
sending just <CR>.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>