Vim: make mouse scroll wheel etc work in iTerm
This works for local sessions running on Mac OS X, and for remote
sessions over SSH (although I had to uncheck and then recheck the
"Enable xterm mouse reporting" checkbox in iTerm to make it happen).

This took some fiddling to get to work under tmux when SSHing into a
remote host; I had to upgrade from tmux 1.4 to 1.5, and add a bunch of
settings to my ~/.tmux.conf.

Signed-off-by: Wincent Colaiuta <win@wincent.com>