tmux/vim: be more insistent about the value of ttymouse
I've found that the VimEnter autocmd is not enough in itself; Vim still
reports "ttymouse=xterm2" but doesn't actual behave as though it could
receive mouse events until you explicitly do a ":set ttymouse=xterm2".
The exact conditions in which this happens are hard to determine, but I
have tried many combinations of opening and closing Vim with and without
file parameters, and also attaching, dettaching, and starting a new
session in tmux.

So far this seems to cover all bases.

Signed-off-by: Wincent Colaiuta <win@wincent.com>