Vim: provide quick override of Solarized colorscheme
For those times when I am using a machine whose Terminal has trouble
dealing with Solarized's colorscheme, provide a means of temporarily
overriding the :color.

When working on such a machine, just touch a file at ~/.vim/dark to fall
back to a reasonable default dark scheme. I chose to do it this way
rather than implementing a callable function because when you're
constantly spawning new editors with `git commit` it's a pain to have to
explicitly call a function in order to make the display look decent.

Signed-off-by: Wincent Colaiuta <win@wincent.com>