Vim/tmux: start doing the right thing with TERM
I used xterm-256color for a long time because one box that I spend a lot
of time on didn't have a terminfo entry for screen-256color. Despite the
fact that the tmux FAQ emphatically warns not to do this, it actually
worked pretty well.

But now that I have the a terminfo entry, let's do the right thing.

Signed-off-by: Wincent Colaiuta <win@wincent.com>