Bash: try to stop prompt corruption issues
I've been observing intermittent prompt corruption issues of late,
regardless of the TERM setting I have inside tmux. Symptoms include
things like:

  - entered text wrapping early, before the full width of the terminal
    has been reached; in this case it behaves as though a carriage
    return had fired, but not a line feed
  - editing command history is made difficult as long lines wrap
    unintelligibly

Notably this only happens when I have a colored prompt. I find the
colored prompt valuable for helping me to be aware of what machine I'm
logged into, so I'd prefer not to drop color.

On the strength of this random suggestion here:

  http://lists.gnu.org/archive/html/bug-bash/2011-02/msg00117.html

I've enclosed those escape sequences "that don't move the cursor within
the terminal" in \[ and \]. It seems to work. (Note: I used to have it
this way until commit 4df9b8110d, when I changed it based on other
examples that I'd seen online. That was a backwards step.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>