Vim: add :Lack, the location-list equivalent of :Ack
Not sure whether I will make much use of this, but may as well try it
out and see how it goes. I'm thinking it might be useful for temporary
searches (ie. when you want a long-lived :Ack result to be there as you
undertake a big refactorig, but you occasionally want to just quickly
check for something else).

Note also that I re-read the help and learnt that the trailing bang on
:grep! (and :lgrep!) is what is suppressing the auto-opening of the
first match. I'm going to leave it as is for now and see if I like it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>