Git: set push.default to "upstream"
The default value of "matching" here is a bit scary, as it makes it too
easy to push things that might not be ready for publication.

Next up is "current", but that could allow you to carelessly create a
branch without meaning to.

The saftest setting is evidently "nothing", which would evidently
require you to aways specify a refspec (either on the command line, or
in the .git/config for the remote, or otherwise implied by options on
the command line like `--all`).

The sweet spot seems to be "upstream", which will push the current
branch to its upstream tracking branch. This should do the right thing
in most cases.

Signed-off-by: Wincent Colaiuta <win@wincent.com>