Git: set branch.master.rebase to true
For the last year now I've been doing `git pull --rebase` whenever I'm
on the master branch. As far as I can tell, this is almost never a bad
thing to do.

Let's make this the default.

Without this setting, you end up having to type `--rebase` out of force
of habit whenever you pull on the master branch. If you forget to type
it, you end up with an unwanted merge which you then need to rewrite
away with `git rebase origin/master`.

With this setting, you'll either get the rebase "for free", or you'll be
warned that you can't rebase because your worktree is dirty. In that
case you eiher have the option of either suppressing the rebase (ie.
with `git pull --no-rebase`) or stashing your changes.

This seems like an acceptable workflow.

Signed-off-by: Wincent Colaiuta <win@wincent.com>