Git: add `br` as an alias for `git-branch`
In general I try not to rely too much on aliases, as I don't want to be
disabled on those occasions when I'm forced to use another person's
environment.

I'm adding this alias, however, as I've noticed that it seems to be a
fairly common one in use among my colleagues, and I want to make them
feel as comfortable as possible when they have to work in my
environment.

Signed-off-by: Wincent Colaiuta <win@wincent.com>