Bash: teach ssh-reagent to kill off dead agents
I occasionally see ssh-agent hang indefinitely while testing candidate
agent sockets. Inspecting them shows that the permissions and ownership
are fine, so it's not clear what the cause of the hang is.

The workaround is to remove the offending socket and try again. This
commit makes that process automatic by getting rid of any socket which
causes `ssh-add -l` to take more than a second to return.

Signed-off-by: Wincent Colaiuta <win@wincent.com>