zsh: get familiar bindings working
Because my EDITOR is vim, looks like the vi bindings are being selected
by default; force emacs bindings to be used instead, seeing as I have
a lot of muscle memory for them.

The one non-emacs binding that I can't live without is the forward and
back history search with the up and down cursor keys.

Signed-off-by: Wincent Colaiuta <win@wincent.com>