Increase priority of /usr/local/bin in PATH
On Mac OS X the default system path puts /usr/local/bin at a
not-very-useful place, after /usr/bin, /bin, /usr/sbin and /sbin, which
means that you can't use it as a place for installing newer versions of
things to override crusty old versions on the system.

So, put an explicit /usr/local/bin before the system path.

On systems with sane default paths (for example, most Linux
distributions that I've tried recently) this means that there'll be an
extra /usr/local/bin early on in the PATH, but it will be harmless.

Signed-off-by: Wincent Colaiuta <win@wincent.com>