bootstrap: handle symlink backups sanely
The backup logic was basically pointless after the first run; the first
run would make a copy of pre-existing files before replacing them with a
symlink, but subsequent runs would make a backup copy of the symlink.
This meant that you would wind up with two symlinks for each dot-file, a
regular link and a ".bak" variant, both pointing at the same file.

So, make backup copies only of things that are actually files or
directories. On subsequent runs, the backup process is omitted, leaving
the original backup intact and ignoring the link.

Signed-off-by: Wincent Colaiuta <win@wincent.com>