Vim: tweak Command-T mappings for opening in a split
I never use <C-p>, <C-n> for their default purposes, instead relying on
<C-j> and <C-k>. So, use <C-p> as a synonym for <C-s>, seeing as the
latter is not working most of the time, despite my `unsetopt
flowcontrol` option in my .zshrc.

(Not sure why, but Zsh often seems to "forget" this option, but after
manually reapplying it all is well, for a while...)

With this mapping I can do <C-p> in a pinch to open something in a split
without having to exit Vim, reset the flowcontrol option, and come back
in.

Signed-off-by: Wincent Colaiuta <win@wincent.com>