Vim: add GitJump command for invoking `git jump` from inside Vim
<leader>d will bring insert `:GitJump diff ` at the command prompt; the
user can either hit enter, add additional arguments (eg. file paths,
diff options), or delete "diff" and specify another mode like "merge".

Not that this required a minor modification to the git-jump script to
detect when STDOUT is connected to a TTY. I'll probably clean that up
and send it upstream.

Signed-off-by: Wincent Colaiuta <win@wincent.com>