Vim: back to Consolas, this time 13
Last time I switched it was because Consolas:14 was looking too wide for
me on the new display. On playing with Consolas:13 I find the width is
no longer a distraction, and the vertical height is more compact that
Monaco:12, meaning I can fit more lines on screen.

So going to give it another try. The fact that "BBEdit" was in the name
was annoying to me, so I grabbed a copy directly from Microsoft as
suggested here:

  http://www.wezm.net/technical/2010/08/howto-install-consolas-font-mac/

The actual download URL is (currently):

  http://www.microsoft.com/en-us/download/details.aspx?id=26342

Signed-off-by: Wincent Colaiuta <win@wincent.com>