Vim: set cursorcolumn
This may be overkill, but I often find that when searching that it is
not always immediately clear which match the cursor is on. When combined
with 'cursorline' this helps that greatly.

If this turns out to be too much I can always do some clever remapping
to turn it on only when I do a search, and turn it off again when I do
`:noh`.

Signed-off-by: Wincent Colaiuta <win@wincent.com>