Git: add "reword" alias
This is a special case of `git amend`, which allows you to modify a
commit message even if you already have additional stages changed in the
index.

Requires Git 1.7.11.3 or later (although the `--only` option was
documented as being useful for this purpose since 1.3, it didn't work as
advertised until 1.7.11.3).

Signed-off-by: Wincent Colaiuta <win@wincent.com>