Vim: use 'cursorcolumn' only during search
As much as I like 'cursorcolumn', the main reason I use it is to make it
clearer which match the cursor is currently on while searching, and it's
a significant performance hit on my machine which has been laboring of
late.

So, temporarily turn it on when initiating a search with `/`, and turn
it off again when cleaning the match highlighting (with `<leader>n`).

I'm commenting out the main declaration for now instead of removing it
entirely, as I am hopeful that one day the performance will be good
enough for me to re-enable it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>