Vim: change cursor shape when in insert mode
Changing the cursor shape will make it more obvious when we are in
insert mode, reducing the likelihood of unintentionally issuing normal
mode commands.

This will work in iTerm 2 on Mac OS X. I tried it in the Terminal.app to
confirm that it degrades gracefully, and it seems to. Alas, it doesn't
work inside Tmux.

Signed-off-by: Wincent Colaiuta <win@wincent.com>