Vim: attempted fix at ambiguous MatchParen highlighting
I find it very hard to see at a glance which of the parens the cursor is
actually on, so make it so that the matched paren is highlighted in a
more subtle fashion.

As noted in the comment, this won't survive a roundtrip to
`background=dark` and back, but as I almost always stay in
`background=light`, this is still a win.

Signed-off-by: Wincent Colaiuta <win@wincent.com>