Git: add comments about compatibility
Sometimes when I land on a new machine with an old version of Git it's
nice to know just how up-to-date I need to be in order to get most (or
all) of the functionality I need.

Rather than going from memory, try to document the minimum required
versions for various aspects of the `.gitconfig` file to work.

As there are a bunch of comments and awkward empty lines in the config
file now, put blank lines between the sections to restore to them a
sense of separation.

Signed-off-by: Wincent Colaiuta <win@wincent.com>