Vim: add `silent!` to custom `/` mapping
This was an attempt to fix a weird insure in which the custom mapping
will move the cursor one column to the right as soon as it is pressed.

In practice, it doesn't work, but neither is the cursor movement problem
unbearable. I am going to leave it as-is for now until I come up with
something better.

Signed-off-by: Wincent Colaiuta <win@wincent.com>