Vim: map `jj` to `<Esc>` in insert mode
It's not often that I need to type "jj" (ironically, this commit message
is one of the few examples), so add it as an insert-mode-only home-row
mapping equivalent to the Escape key.

This is going to be a huge amount of muscle memory to overcome, but I am
going to try running with it for a while. I'm expecting my initial
attempt at using it will consist of:

  - type in insert mode for a while
  - hit `jj`
  - hit `<Esc>` anyway, as a kind of nervous tic

Signed-off-by: Wincent Colaiuta <win@wincent.com>