Vim/Tmux: hack fix for cursor-related display bugs
I'm not entirely happy with this solution to my display corruption
problems, because I don't fully understand what is causing the problem
nor how the solution works. I'm going to trial it anyway, and I've cut
down the terminal overrides to what appears to be the most minimal set
capable of fixing the issue.

Signed-off-by: Wincent Colaiuta <win@wincent.com>