Git: add "lg" alias
"oneline" is still my go-to workhorse, but adding "lg" as a variant
which provides richer metadata (relative timestamps, authors) for those
times when I want to get a sense of who has been doing what and when in
a project.

Not that this couldn't replace "oneline", even if I wanted it to,
because it annoyingly uses the pager, even when one isn't needed (eg.
`git lg -10` dumps you into a pager). If I could figure out a way to
work around that I might consider switching.

Signed-off-by: Wincent Colaiuta <win@wincent.com>