Vim: use bracketed paste mode
This works on my (OS X) laptop inside and outside of tmux.

This is somewhat black-magical, but pleasingly works even in edge-casey
scenarios like alternating pastes between Vim and the shell (in
different tmux panes; or by foregrounding and backgrounding the editor
process).

On CentOS, looks like it only works outside tmux. This might be because
the CentOS box I'm using is using tmux 1.6 while my the Mac has 1.7, so
I'll be trying upgrading if I can solve the build issues I had when
trying to update to 1.7 previously.

The tmux 1.7 changelog does report that "bracketed paste mode" support
was added:

  http://sourceforge.net/projects/tmux/files/tmux/tmux-1.7/

but the other references I found refer to the addition of a  `-b` option
to the tmux `paste-buffer` command, which isn't at all relevant here:

  http://www.mail-archive.com/tmux-cvs@lists.sourceforge.net/msg00572.html

Signed-off-by: Wincent Colaiuta <win@wincent.com>