tmux: add rate-limiting
This is one of the main features that made me want to start using tmux
1.7: the ability to throttle updates when there's a "runaway" process
hammering the terminal with too many updates.

The default values of 250 triggers/millisecond and 100ms between updates
work ok for a test run of `yes`. I found that when `cat`-ing a log file
I needed to tweak the numbers a little more.

Clearly I'll have to fine-tune this further once I'm plugged into the
external monitor at work, but for now this is an awesome improvement.

Signed-off-by: Wincent Colaiuta <win@wincent.com>