*sh: update setting of JAVA_HOME
We still have a somewhat unsightly hard-coded value for use on my CentOS
box, but I don't often use Java there so I don't mind too much.

Another slightly irksome thing is that there is a subtle dependency
between the .shells/path and .shells/exports file: the former depends on
the latter (it needs EC2_HOME to be defined), but ideally the latter
would make use of the former (as noted in the comment). For now I have
no bright ideas how to make that better.

Signed-off-by: Wincent Colaiuta <win@wincent.com>