Vim: ditch Align.vim for Tabular
Align.vim works pretty well for aligning Ruby 1.8-style hashes (eg. with
`:Align =>`:

  # this:               becomes:
  :foo => 'a',          :foo  => 'a',
  :bing => 'b',         :bing => 'b',

or JS objects (eg. with `:Align :`):

  // this:              becomes:
  foo: 1,               foo  : 1,
  bing: 2,              bing : 2,

but I can't for the life of me figure out how to get it to deal with
Ruby 1.9+ hash syntax; ie:

  // from this:         to:
  foo: 1,               foo:  1,
  bing: 2,              bing: 2,

Tabular can handle these examples as follows:

  :Tab /=>
  :Tab /:
  :Tab /:\zs

(where `\zs` is a magical zero-width atom provided by Vim that sets the
start of a match; see `:h \zs`.)

Of course, both Tabular (and Align) can do much, much more than this,
but this is the stuff I typically use it for.

Related:

  http://vimcasts.org/episodes/aligning-text-with-tabular-vim/

Signed-off-by: Wincent Colaiuta <win@wincent.com>