*sh: add `groot` alias
I would really have liked this to be a Git alias (for example, so I
could type `git root` to go to the root of a Git repo).

Unfortunately, I can't see any way to do that. Git runs aliases in
subshells, so changing dirctories inside them has no effect. This is
true regardless of whether we implement the alias in terms of a
function, a shell command, or a separate executable.

So, this is the next best thing.

`groot` takes you to the top level of a repo, if you're in one, and
leaves you where you are otherwise.

Signed-off-by: Wincent Colaiuta <win@wincent.com>