Git: teach `fixup` alias to take additional params
Make use of magical `${@:2}` variable to pass parameters 2 through N to
the command.

This means that in addition to:

  git fixup HEAD~

you can do things like:

  git fixup HEAD~ --no-verify

There is a minor wart here which is that normally options would go
before non-option arguments; ie:

  git fixup --no-verify HEAD~

but switching the order of the alias would break the base (optionless)
case.

For now I'll run with this and if it gets annoying will bite the bullet
and write a proper shell function.

Signed-off-by: Wincent Colaiuta <win@wincent.com>