Zsh: weaken compinit security checks for the benefit of Homebrew
For security, I run as an unprivileged user, and have a separate admin
account for admin purposes.

Alas, by an unfortunate combination:

  - Homebrew wants to be installed as an admin user
  - Zsh wants to check that the ownership and permission on compinit
    files

This means that Zsh will complain on boot, because the files will be
owned by another user who isn't root. This commit switches off the
security check.

Sigh. It was either that, or run as an admin user, which isn't
desirable either. (Or, dispense with Homebrew...)

So here I am choosing convenience over security, because I've set up too
many machines in my life and I want to streamline the process.

Signed-off-by: Wincent Colaiuta <win@wincent.com>