Vim: <leader>r cycles through number modes
Previously <leader>r switched between relative and absolute line
numbering. Now it cycles through relative, absolute and no line
numbering.

This is based on a super-cute trick stolen from:

 http://www.reddit.com/r/vim/comments/19u4u0/what_normal_mode_mappings_do_you_have_in_your/c8rh8ox

`<c-r>=` inserts the result of an expression at the cursor, which in
this case is a mapping from the current states of the `'rnu'` and `'nu'`
settings to the next desired state:

  - rnu off (0), nu off (0) -> rnu on,  nu off (`:set rnu`)
  - rnu off (0), nu on  (1) -> rnu off, nu off (`:set nonu`)
  - rnu on  (1), nu off (0) -> rnu off, nu on  (`:set nu`)

For more info, see:

  http://stackoverflow.com/questions/10862457/what-does-c-r-means-in-vim

Signed-off-by: Wincent Colaiuta <win@wincent.com>