Vim: remove ttymouse hack
For some reason that I never fully understood, the `set ttymouse`
directive wasn't sticking, so I had to keep resetting it from an
autocmd. Perhaps it was an interaction with some plug-in I was using, or
something that only manifesting under other versions of iTerm or tmux or
Vim.

Whatever the cause, it no longer seems to be an issue, so remove the
hack.

Signed-off-by: Wincent Colaiuta <win@wincent.com>