Vim: add hack for indicating iTerm-ness on remote hosts
At my work, our sshd is locked down to with PermitUserEnvironment at the
default value of false. Furthermore, the list of AllowEnv variables is
also locked down fairly tight (LANG, LC_* and such).

This means that ITERM_PROFILE, the way that the .vimrc is detecting that
it's running inside iTerm 2, is not getting passed along.

I'll see if I can get the sshd_config modified at work, but in the
meantime, rather than abusing one of the permitted variables to pass
along the hint that iTerm is running, I'm adding this other hack; a
"dot" file that I can touch whenever I want Vim to know that iTerm is on
the scene.

Signed-off-by: Wincent Colaiuta <win@wincent.com>