Vim: move Gundo configuration out of ~/.vimrc
With this and other similar changes I am somewhat abusing the intent of
the "after" directory. The intent here is really to override defaults
provided by plug-ins, especially overrides that can't be done at
~/.vimrc evaluaton time.

Nevertheless, it is enabling me to clean out a very cluttered ~/.vimrc
and turn it into something more palatable.

Signed-off-by: Wincent Colaiuta <win@wincent.com>