Switch back to Monaco 12-pt
I've just learned that the baseline on Consolas is incorrect:

  http://mbauman.net/geek/2009/03/15/minor-truetype-font-editing-on-a-mac/

and now I can't stop noticing it.

Apparently, the old version shipped with BBEdit 9.1 is right, but that
one is missing the bold and italics variants:

  http://www.command-tab.com/2008/12/16/consolas-cursor-fix/

Oddly, it displays fine in the Terminal.app, so Apple must be doing
something to compensate.

If I can get a way to run Apple's font tools to fix the font I will, but
for now it seems I can't as they are PPC-only and I can't see an updated
version on their developer downloads page (not sure if it is busted for
everyone or just for me, but I can't see any downloads at all).

Signed-off-by: Wincent Colaiuta <win@wincent.com>