Zsh: fix warning in tmux wrapper function
Using the `[`/`test` built-in here is producing a warning like the
following every time I pass args to the wrapper function:

  tmux:[:1: too many arguments

The syntax appears to work fine in an interactive shell, so it's
possible that there is an environmental explanation, or a load-order
explanation.

Either way, switch the built-in conditional expression syntax. This
works in both Bash and Zsh without warnings.

Signed-off-by: Wincent Colaiuta <win@wincent.com>