Zsh: improve SPROMPT
Replace the standard:

  zsh: correct 'lss' to 'less' [nyae]?

with:

  zsh: correct 'lss' to 'less' [yes, no, edit, abort]?

where the faulty input and the proposed correction appear in red, and
the first letter of each of "yes" ,"no", "edit" and "abort" is bolded
and underlined.

Note that I changed the ordering in the default prompt from nyae to
ynea, to better match the frequency with which I typically select from
among those options.

Signed-off-by: Wincent Colaiuta <win@wincent.com>