Zsh: put Git info into RPROMPT
I've always valued a terse PROMPT, and preferably a one-line one, so I
don't really want to go for something lengthy, elaborate, or even
something spread over two or more lines as seems to be all the rage
nowadays.

Using vcs_info and RPROMPT enables me to jam a fair bit of useful info
off on the right without it getting in the way.

I've included a link to the documentation in a comment as it can be had
to find.

This replaces the old .git-prompt.sh script.

Signed-off-by: Wincent Colaiuta <win@wincent.com>