*sh: set TERMINFO if appropriate
On my OS X box I'm running a mostly-vanilla set up, where "vanilla" is
defined as letting Homebrew install everything. On my CentOS box, in
contrast, I had to build a statically-linked tmux in order to get things
working with tmux 1.7, 1.8:

  https://wincent.com/wiki/Building_tmux_1.7_on_CentOS_5.8
  https://wincent.com/wiki/Building_tmux_1.8_on_CentOS_5.9

This means I have a statically-linked ncurses library, and it put a
bunch of terminfo files at $HOME/share/terminfo.

In order to be able to use italics in the terminal and Vim on this box,
I need to be able to install terminfo files there and the shell needs to
be able to find them. So, if that directory is present, set the TERMINFO
variable.

Signed-off-by: Wincent Colaiuta <win@wincent.com>