tmux: add back binding to copy last-copied text to Clipper
With the addition of the copy-pipe command to tmux 1.8 I though that my
old `<prefix> <c-y>` binding wasn't going to be useful any more.

It turns out that there is still a use case for it: when you use the
mouse to enter copy mode and select text, it gets copied, but not to
Clipper.

Add back the binding, but with a tweak: make it `<prefix> y` seeing as
that mapping is available, easier to type, and closer to the `<leader>
y` mapping that has the same purpose in my Vim configuration.

Signed-off-by: Wincent Colaiuta <win@wincent.com>