Sh*: overwrite old terminfo entries
After a few days of testing this is stable enough to "merge". This means
I can connect to other hosts and won't have to worry about warnings
that the "terminal is not fully functional" (because the remote hosts
naturally don't recognize my screen-256color-italic TERM type etc).

Added instructions on how to backup and sanity-check the generated
files.

Signed-off-by: Wincent Colaiuta <win@wincent.com>