Git: make "lg" the new "oneline"
As the comment indicated, I was using "lg" supplementally to "oneline"
rather than as a replacement for it.

I just re-tested it, and some recent version of Git has fixed the
annoying problem with the pager.

This is great, as it means I can make "lg" the new "oneline", and there
is no longer any "lg" alias (which was a little annoying in things like
Rails projects which have a "log" directory, and Zsh would constantly
offer to "autocorrect" my errant "lg" alias).

Signed-off-by: Wincent Colaiuta <win@wincent.com>