Vim: remove jk and kj mappings
In my trial of jj, jk and kj, jk is losing out. I can't type it as
comfortably as either of the other two. It feels downright awkward. So,
I'm removing it.

kj feels pretty good to type, but somehow my muscle memory has decided
that jj is going to be the candidate of choice. So jj it is.

I reordered a couple of things in making this change in order to
preserve the desired sectional grouping within the file.

Signed-off-by: Wincent Colaiuta <win@wincent.com>