Vim: conclude supertab experiment
I'm afraid this is too annoying to leave in place. For example, given
code like:

  var foobar = 1,
      baz|

(cursor position indicated with "|"), I'd like to be able to hit tab and
have it move the cursor up to:

  var foobar = 1,
      baz    |

so I can type:

  var foobar = 1,
      baz    = 2;

If there is already a completion for "baz" elsewhere however, it just
sits there.

I'll stick with CTRL-N, which I have pretty well ingrained into the
muscle memory by now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>