Vim: override standard JavaScript "fun" snippet
I find the standard "fun" snippet to be a little annoying as by far the
most common function type I declare is anonymous (often assigned to a
property), so I end up having to awkwardly delete the function name and
the preceding space.

This snippet includes the preceding space in the placeholder, so it can
be deleted with a single tap of the delete key.

Signed-off-by: Wincent Colaiuta <win@wincent.com>