Vim: use augroups
As suggested by @lencioni here:

  https://github.com/wincent/wincent/commit/31a7b0a#commitcomment-4987362

This is a bit of a non-problem for me, as I basically never reload my
.vimrc, but might as well do things "The Right Way".

I've "namespaced" the augroup names by prefixing them with "wincent_",
to avoid potential clashes with groups that might be defined by
plug-ins.

I haven't touched the auto commands in ftdetect files because Vim
already handles those for us by sticking them in a group and clearing it
on any reload.

Signed-off-by: Wincent Colaiuta <win@wincent.com>